ISSUE #2490: Use version assets by app version instead of a random UUID
